Add missing imports and refactor androidTest directory
