+++ LocalPayment/src/androidTest/java/com/braintreepayments/api/localpayment/LocalPaymentClientTest.java
@@ -1,15 +1,13 @@
-package com.braintreepayments.api;
+package com.braintreepayments.api.localpayment;
 
 import static junit.framework.Assert.assertNotNull;
-
 import static org.junit.Assert.assertTrue;
 
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.internal.runner.junit4.AndroidJUnit4ClassRunner;
 
-import com.braintreepayments.api.localpayment.LocalPaymentAuthRequest;
-import com.braintreepayments.api.localpayment.LocalPaymentClient;
-import com.braintreepayments.api.localpayment.LocalPaymentRequest;
+import com.braintreepayments.api.core.BraintreeClient;
+import com.braintreepayments.api.core.PostalAddress;
 
 import org.junit.Before;
 import org.junit.Test;
