+++ LocalPayment/src/test/java/com/braintreepayments/api/localpayment/LocalPaymentAnalyticsUnitTest.kt
@@ -1,6 +1,5 @@
 package com.braintreepayments.api.localpayment
 
-import com.braintreepayments.api.localpayment.LocalPaymentAnalytics
 import org.junit.Assert.assertEquals
 import org.junit.Test
